gh-4700 Fix proxy-dev.yml and expected.yml

+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/handler/ForwardHttpPostDestination.java
@@ -400,7 +400,7 @@ public class ForwardHttpPostDestination implements ForwardDestination {
                 0L);
         final int attempts = NullSafe.getOrElse(retryState, RetryState::attempts, -1);
         final StroomDuration retryDelay = forwardHttpPostConfig.getRetryDelay();
-        final int retryDelayGrowthFactor = forwardHttpPostConfig.getRetryDelayGrowthFactor();
+        final double retryDelayGrowthFactor = forwardHttpPostConfig.getRetryDelayGrowthFactor();
         final long retryDelayMs;
         if (retryDelayGrowthFactor > 1 && retryState != null) {
             retryDelayMs = Math.min(
